lambdas/typedb_search_query/requirements.txt
typedb-client==2.12.1
inflect==6.0.2